infrastructure/requirements.txt
aws-cdk.core==1.192.0
aws-cdk.aws_s3==1.192.0
aws-cdk.aws_lambda==1.192.0
aws-cdk.aws_ecr==1.192.0
aws-cdk.aws_ec2==1.192.0